# runtime.Callers: use map to reduce allocations and performance improv…