goos: darwin
goarch: amd64
pkg: runtime
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
benchmark                      old ns/op     new ns/op     delta
BenchmarkCaller-12             524           238           -54.52%
BenchmarkCallerParallel-12     114           60.5          -46.80%

benchmark                      old allocs     new allocs     delta
BenchmarkCaller-12             2              0              -100.00%
BenchmarkCallerParallel-12     2              0              -100.00%

benchmark                      old bytes     new bytes     delta
BenchmarkCaller-12             216           0             -100.00%
BenchmarkCallerParallel-12     216           0             -100.00%